Don't use file://, do require absolute paths
xdg-open documents itself as working on ordinary files as well as
file:// URLs, and in practice does not seem to necessarily *work* on
file:// URLs in the first place.

I may yet change my mind, but I don't really want to try and define
and/or document tkrunner's working directory; it's probably best to ban
relative paths.